rename Sql namespaces to Scripts
