+++ app/Library/Library.csproj
@@ -53,20 +53,20 @@
     <Compile Include="Releases\Tasks\Bump.cs" />
     <Compile Include="Command.cs" />
     <Compile Include="Please.cs" />
-    <Compile Include="Sql\Script.cs" />
-    <Compile Include="Sql\RunException.cs" />
-    <Compile Include="Sql\Tasks\CheckForVersionTable.cs" />
-    <Compile Include="Sql\Tasks\RunSql.cs" />
-    <Compile Include="Sql\Tasks\InsertInstalledVersion.cs" />
-    <Compile Include="Sql\Tasks\Run.cs" />
-    <Compile Include="Sql\Tasks\RunProcess.cs" />
-    <Compile Include="Sql\Tasks\SplitSqlOnGo.cs" />
-    <Compile Include="Sql\Version.cs" />
-    <Compile Include="Sql\Tasks\FetchInstalledVersions.cs" />
-    <Compile Include="Sql\Tasks\CreateVersionTable.cs" />
-    <Compile Include="Sql\Tasks\GetScripts.cs" />
-    <Compile Include="Sql\Tasks\RunScripts.cs" />
-    <Compile Include="Sql\Tasks\RunMissingVersions.cs" />
+    <Compile Include="Scripts\Script.cs" />
+    <Compile Include="Scripts\RunException.cs" />
+    <Compile Include="Scripts\Tasks\CheckForVersionTable.cs" />
+    <Compile Include="Scripts\Tasks\RunSql.cs" />
+    <Compile Include="Scripts\Tasks\InsertInstalledVersion.cs" />
+    <Compile Include="Scripts\Tasks\Run.cs" />
+    <Compile Include="Scripts\Tasks\RunProcess.cs" />
+    <Compile Include="Scripts\Tasks\SplitSqlOnGo.cs" />
+    <Compile Include="Scripts\Version.cs" />
+    <Compile Include="Scripts\Tasks\FetchInstalledVersions.cs" />
+    <Compile Include="Scripts\Tasks\CreateVersionTable.cs" />
+    <Compile Include="Scripts\Tasks\GetScripts.cs" />
+    <Compile Include="Scripts\Tasks\RunScripts.cs" />
+    <Compile Include="Scripts\Tasks\RunMissingVersions.cs" />
     <Compile Include="Commands.cs" />
     <Compile Include="ICommand.cs" />
     <Compile Include="Option.cs" />
