+++ app/Library/Scripts/Tasks/CheckForVersionTable.cs
@@ -2,7 +2,7 @@
 using Simpler;
 using Simpler.Data;
 
-namespace Library.Sql.Tasks
+namespace Library.Scripts.Tasks
 {
     public class CheckForVersionTable : InOutTask<CheckForVersionTable.Input, CheckForVersionTable.Output>
     {
