+++ app/Tests/App.config
@@ -1,7 +1,7 @@
 ﻿<?xml version="1.0" encoding="utf-8" ?>
 <configuration>
   <connectionStrings>
-    <add name="Test" connectionString="Data Source=Sql\files\test.db" providerName="System.Data.SQLite" />
+    <add name="Test" connectionString="Data Source=Scripts\files\test.db" providerName="System.Data.SQLite" />
   </connectionStrings>
   <appSettings>
     <add key="RunSqlTimeout" value="60"/>
