+++ app/Tests/Scripts/files/sql/repeatable/create-four-tables.sql
